requirements.txt
torch==2.1.2
pytorch_lightning==1.9.0
transformers==4.36.0
umap-learn==0.5.3
networkx==2.8.6
matplotlib==3.6.0
einops==0.6.1
bitsandbytes==0.42.0
diffusers==0.23.1
hydra-core==1.3.2
hvplot==0.8.3
cloudpathlib==0.15.0
s-dbw==0.4.0
accelerate==0.24.1
ray==2.32.0
piq==0.7.1
rich==13.5.2
omegaconf==2.3.0
opencv-python==4.10.0.84
scikit-image==0.21.0
natsort==8.4.0
datasets==2.16.1
pyarrow>=14.0.0
xformers==0.0.23.post1
numpy==1.25.2
inflect==7.3.1
